Pull FIPS collection into execution environment
In order to make a more turnkey solution for running opcap-ansible,
this adds the fips-assessment collection to the execution environment.

The GitHub action uses an ansible.cfg from a secret, since it has
both the URL for the private automation hub and the token for the
hub.

Ansible.cfg is moved to ansible.cfg.template and added to .gitignore
to hopefully prevent tokens from leaking.

Fixes #98

Signed-off-by: Brad P. Crochet <brad@redhat.com>